final group report + individual report upload
