+++ Individual-Final-Report/Sanchit-Vijay/Code/run.sh
@@ -1,0 +1,39 @@
+#!/bin/bash
+
+# Navigate to the directory containing the first main.py
+#cd /path/to/first/directory
+echo "Running first main1.py"
+python3 main_ft.py
+
+# Navigate to the directory containing the second main.py
+#cd /path/to/second/directory
+echo "Running second main2.py"
+python3 main2.py
+
+echo "Running third main3.py"
+python3 main3.py
+
+#echo "Running fourth main3.py"
+python3 main4.py
+
+python3 main5.py
+
+python3 main6.py
+
+python3 main7.py
+
+python3 main8.py
+
+python3 main9.py
+
+python3 main10.py
+
+python3 main11.py
+
+python3 main12.py
+# Add similar blocks for additional main.py files
+
+# Return to the original directory if needed
+#cd -
+
+echo "All scripts executed successfully."
